+++ packages/annotation/tests/symbols/InvertedTriangle.test.tsx
@@ -11,26 +11,26 @@ describe('InvertedTriangle', () => {
             </Chart>
         )
         const result = screen.getByRole('chart-content').querySelector('polygon')
-        expect(result?.getAttribute('role')).toContain('default')
+        expect(result).toBeDefined()
     })
 
-    it('creates a default inverted triangle without role', () => {
+    it('creates an inverted triangle with role', () => {
         render(
             <Chart {...chartProps}>
-                <InvertedTriangle setRole={false} />
+                <InvertedTriangle variant={'custom-triangle'} />
+            </Chart>
+        )
+        const result = screen.getByRole('chart-content').querySelector('polygon')
+        expect(result?.getAttribute('role')).toBe('custom-triangle')
+    })
+
+    it('creates an inverted triangle without role', () => {
+        render(
+            <Chart {...chartProps}>
+                <InvertedTriangle variant={'custom-triangle'} setRole={false} />
             </Chart>
         )
         const result = screen.getByRole('chart-content').querySelector('polygon')
         expect(result?.getAttribute('role')).toBeNull()
-    })
-
-    it('creates a custom inverted triangle', () => {
-        render(
-            <Chart {...chartProps}>
-                <InvertedTriangle variant={'test'} cx={10} cy={20} r={5} setRole={true} />
-            </Chart>
-        )
-        const result = screen.getByRole('test')
-        expect(result.getAttribute('role')).toContain('test')
     })
 })