+++ packages/core/tests/shapes/Line.test.tsx
@@ -9,21 +9,33 @@ describe('Line', () => {
                 <Line x1={0} y1={0} x2={10} y2={20} />
             </Chart>
         )
-        const result = screen.getByRole('default')
+        const result = screen.getByRole('chart-content').querySelector('line')
         expect(result?.getAttribute('y1')).toContain('0')
         expect(result?.getAttribute('x2')).toContain('10')
         expect(result?.getAttribute('y2')).toContain('20')
-        expect(result?.getAttribute('role')).toContain('default')
+        expect(result?.getAttribute('role')).toBeNull()
     })
 
-    it('creates a default line without role', () => {
+    it('creates a line variant with role', () => {
         render(
             <Chart {...chartProps}>
-                <Line x1={0} y1={0} x2={10} y2={20} setRole={false} />
+                <Line variant={'custom-line'} x1={0} y1={0} x2={10} y2={20} setRole={true} />
             </Chart>
         )
-        const result = screen.getByRole('chart-content')
-        expect(result.querySelectorAll('line')).toHaveLength(1)
+        const result = screen.getByRole('chart-content').querySelector('line')
+        expect(result?.getAttribute('class')).toEqual('customLine')
+        expect(result?.getAttribute('role')).toEqual('custom-line')
+    })
+
+    it('creates a line variant without role', () => {
+        render(
+            <Chart {...chartProps}>
+                <Line variant={'custom-line'} x1={0} y1={0} x2={10} y2={20} setRole={false} />
+            </Chart>
+        )
+        const result = screen.getByRole('chart-content').querySelector('line')
+        expect(result?.getAttribute('class')).toEqual('customLine')
+        expect(result?.getAttribute('role')).toBeNull()
     })
 
     it('assigns markers', () => {